refactor(output): `LottoResultView` -> `LottosView`
